Mark simple and moderate captchas as deprecated.
Change extreme captcha to use random colours, instead of fixed ones.
Change it to have lines to every character: the characters in the solution
are now connected by dotted lines; characters not in the solution are
connected by solid lines. Replace dots with polygons.

Add new captchas with a TrueType font. The solution is obscured by small
characters or by rectangles, ellipses, slices of ellipses, and polygons.
These shapes are either filled or outlined.

Fixes #13